Custom inspect methods for Trip and TripDispatcher
These are necessary to prevent Ruby from entering an infinite loop when
trying to inspect an instance of these classes (due to the circular
references between Trip and Driver/Passenger).

Technically it's not required to have a custom inspect for
TripDispatcher, but the default behavior would result in printing out
every single Trip, Driver, and Passenger in its instance variables which
is a huuuuge block of text that is entirely unhelpful.

Instead, we list the counts of each of those arrays.